Create /pokemon URL
